+++ demos/components/vectorstore/vs_utilities.py
@@ -21,6 +21,7 @@ def sanitize_string(some_text):
     cleaner_name = cleaner_name.replace(" ", "_")  # spaces to underscores
     cleaner_name = re.sub(r'[^a-zA-Z0-9_-]', '-', cleaner_name)  # replace invalid characters with spaces
     cleaner_name = cleaner_name.rstrip('-')  # remove trailing dashes
+    cleaner_name = cleaner_name.rstrip('_')  # remove trailing underscores
     return cleaner_name
