+++ demos/rag/launch_upload_ui.py
@@ -1,7 +1,7 @@
 import gradio as gr
 from tqdm import tqdm
 
-from demos.components.vectorstore.cdb_utilities import sanitize_filename
+from demos.components.vectorstore.vs_utilities import sanitize_filename
 from demos.components.vectorstore.chroma_document_store import ChromaDocumentStore
 
 cdb_store = ChromaDocumentStore(path='store/')
